+++ aisearch/src/main/java/com/elastic/aisearch/service/HistoryService.java
@@ -1,10 +1,13 @@
 package com.elastic.aisearch.service;
 
+import com.elastic.aisearch.dto.Messages.FailMessageDTO;
+import com.elastic.aisearch.dto.Messages.SuccessMessageDTO;
 import com.elastic.aisearch.entity.History;
 import com.elastic.aisearch.repository.HistoryRepository;
 
 import lombok.AllArgsConstructor;
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,7 +29,11 @@ public class HistoryService {
         historyRepository.deleteAllByUserId(userId);
     }
 
-    public void deleteHistory(Integer id) {
-        historyRepository.deleteById(id);
+    public ResponseEntity<?> deleteHistory(Integer id, Integer userId) {
+        if (historyRepository.verifyIdOwner(id, userId)) {
+            historyRepository.deleteById(id);
+            return ResponseEntity.ok().body(new SuccessMessageDTO("success_delete"));
+        }
+        return ResponseEntity.ok().body(new FailMessageDTO("failed_delete_not_owner"));
     }
 }
